Add Updated Script and Executable File
This is an updated version of the spreadsheet corrector that opens an Excel file included in the same directory as the python script. I have also converted this into an .exe executable file using pyinstaller. Again, simply place the Excel file that you want to convert in the same folder as the .exe file.